use of 32 bit version

+++ utils/hashmurmur_main.cpp
@@ -7,9 +7,8 @@ int main(int argc, char** argv)
 {
   std::string line;
 
-  utils::hashmurmur<uint64_t> hasher;
+  utils::hashmurmur<uint32_t> hasher;
   
   while (std::getline(std::cin, line))
     std::cout << hasher(line.begin(), line.end(), 0) << std::endl;
-    
 }
